seperate the sensorUpdateTask
